Split value absent metadata tests
